+++ python/cugraph/link_analysis/pagerank_wrapper.pyx
@@ -27,6 +27,7 @@ import cudf
 import cudf._lib as libcudf
 import rmm
 import numpy as np
+import numpy.ctypeslib as ctypeslib
 
 
 def pagerank(input_graph, alpha=0.85, personalization=None, max_iter=100, tol=1.0e-5, nstart=None):
@@ -62,25 +63,46 @@ def pagerank(input_graph, alpha=0.85, personalization=None, max_iter=100, tol=1.
 
     cdef bool has_guess = <bool> 0
     if nstart is not None:
-        df['pagerank'][nstart['vertex']] = nstart['values']
+        if len(nstart) != num_verts:
+            raise ValueError('nstart must have initial guess for all vertices')
+        if input_graph.renumbered is True:
+            renumber_series = cudf.Series(input_graph.edgelist.renumber_map.index,
+                                          index=input_graph.edgelist.renumber_map, dtype=np.int32)        
+            vertex_renumbered = renumber_series.loc[nstart['vertex']]
+            df['pagerank'][vertex_renumbered] = nstart['values']
+        else:
+            df['pagerank'][nstart['vertex']] = nstart['values']
         has_guess = <bool> 1
 
-    cdef gdf_column c_identifier_col = get_gdf_column_view(df['vertex'])
-    cdef gdf_column c_pagerank_col = get_gdf_column_view(df['pagerank'])
-    cdef gdf_column c_pers_vtx
-    cdef gdf_column c_pers_val
+    #TODO FIX ME when graph class is upgraded to remove gdf_column
+    cdef gdf_column c_identifier = get_gdf_column_view(df['vertex'])
 
-    g.transposedAdjList.get_vertex_identifiers(&c_identifier_col)
+    cdef uintptr_t c_pagerank_val = get_column_data_ptr(df['pagerank']._column)
+    cdef uintptr_t c_pers_vtx = <uintptr_t>NULL
+    cdef uintptr_t c_pers_val = <uintptr_t>NULL
+    cdef sz = 0
+
+    if personalization is not None:
+        sz = personalization['vertex'].shape[0]
+        if input_graph.renumbered is True:
+            renumber_series = cudf.Series(input_graph.edgelist.renumber_map.index,
+                                          index=input_graph.edgelist.renumber_map, dtype=np.int32)
+            vertex_renumbered = renumber_series.loc[personalization['vertex']]
+            c_pers_vtx = get_column_data_ptr(vertex_renumbered._column)
+        else:
+            personalization_vertex = personalization['vertex'].astype(np.int32)
+            c_pers_vtx = get_column_data_ptr(personalization_vertex._column)
+        personalization_values = personalization['values'].astype(df['pagerank'].dtype)
+        c_pers_val = get_column_data_ptr(personalization_values._column)
     
-    if personalization is None:
-        c_pagerank.pagerank(g, &c_pagerank_col, <gdf_column*> NULL, <gdf_column*> NULL,
-                <float> alpha, <float> tol, <int> max_iter, has_guess)
-    else:
-        c_pers_vtx = get_gdf_column_view(personalization['vertex'])
-        c_pers_val = get_gdf_column_view(personalization['values'])
-        c_pagerank.pagerank(g, &c_pagerank_col, &c_pers_vtx, &c_pers_val,
-                <float> alpha, <float> tol, <int> max_iter, has_guess)
+    if (df['pagerank'].dtype == np.float32): 
+        c_pagerank.pagerank[int, float](g, <float*> c_pagerank_val, sz, <int*> c_pers_vtx, <float*> c_pers_val,
+                                     <float> alpha, <float> tol, <int> max_iter, has_guess)
+    else: 
+        c_pagerank.pagerank[int, double](g, <double*> c_pagerank_val, sz, <int*> c_pers_vtx, <double*> c_pers_val,
+                            <float> alpha, <float> tol, <int> max_iter, has_guess)
 
+    g.transposedAdjList.get_vertex_identifiers(&c_identifier)
     if input_graph.renumbered:
         df['vertex'] = input_graph.edgelist.renumber_map[df['vertex']]
     return df